+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## v1.0.0-rc1
 
-- Prepared public release candidate for JBD submission.
+- Prepared public release candidate for manuscript submission.